+++ src/Resources/contao/languages/de/tl_page.php
@@ -8,11 +8,14 @@
 $lang = &$GLOBALS['TL_LANG']['tl_page'];
 
 /**
- * fields
+ * Fields
  */
+$lang['addHeadDefaultImage'] = ['Fallback-Bild für die Meta-Tags hinzufügen', 'Dieses Bild wird bspw. für og:image und twitter:image verwendet.'];
+$lang['headDefaultImage'] = ['Fallback-Bild (mind. 200x200)', 'Wählen Sie hier ein Bild aus. Nur PNG- und JPG-Dateien sind erlaubt.'];
 $lang['twitterSite'] = ['Twitter @username', 'Der Twitter @username der einer Twitter-Karte zugewiesen werden soll (twitter:site Attribut).'];
 
 /**
- * legends
+ * Legends
  */
-$lang['twitter_legend'] = 'Twitter';
+$lang['head_legend'] = 'Head-Bundle';
+$lang['head_twitter_legend'] = 'Head-Bundle - Twitter';