+++ src/Resources/contao/languages/en/tl_page.php
@@ -8,11 +8,14 @@
 $lang = &$GLOBALS['TL_LANG']['tl_page'];
 
 /**
- * fields
+ * Fields
  */
+$lang['addHeadDefaultImage'] = ['Add fallback image for the meta tags', 'This image is used for og:image and twitter:image.'];
+$lang['headDefaultImage'] = ['Fallback image (at least 200x200)', 'Choose an image here. Only PNG and JPG files allowed.'];
 $lang['twitterSite'] = ['Twitter @username', 'The Twitter @username a twitter card should be attributed to (twitter:site attribute).'];
 
 /**
- * legends
+ * Legends
  */
-$lang['twitter_legend'] = 'Twitter';
+$lang['head_legend'] = 'Head-Bundle';
+$lang['head_twitter_legend'] = 'Head-Bundle - Twitter';